@@ -25,7 +25,7 @@
     "font-awesome": "^4.7.0",
     "google-map-react": "^2.1.9",
     "moment": "^2.28.0",
-    "react": "^16.13.1",
+    "react": "^16.14.0",
     "react-accessible-accordion": "^3.3.5",
     "react-animated-text-builders": "^1.1.0",
     "react-apexcharts": "^1.3.9",
